multi: Only use module-scoped data in tests.
This updates the tests in database/ffldb and blockchain/stake to
reference the block tests within the same module as opposed to escaping
the module to find it in the blockchain module.

This is necessary for the new test all functionality of the upcoming
go1.11 release since it individually tests each module in a separate
path and therefore it prevent the ability to use relative paths that
escape the module.

Unfortunately, this means the data had to be copied, but, but it's not a
huge amount of data, and it seems that git mostly deduplicates the data
internally.