blockchain: Reduce block availability assumptions.
This modifies the code in blockchain to make less assumptions about the
availability of the block data by checking the status flags to determine
whether or not it is available.

While currently there aren't actually ever any nodes in the block index
that don't have the full block data available, that will not always be
true in the future.