rpcserver: Remove createrawssgen RPC.
This removes the handler for the createrawssgen RPC and deprecates the
associated dcrjson structures.

This is being done because it doesn't make any sense for votes to be
created over the RPC interface.  In order for voting on the previous
block to have any semblance of relevance, the software producing the
vote clearly must have code to deserialize and understand the relevant
data structures.