blockchain: Separate tx input stake checks.
This separates all of the validation checks in CheckTransactionInputs
related to each allowed type of stake transaction into separate
functions dedicated to each type.

It also renames the utxo view parameter to CheckTransactionInputs to be
consistent with the name used in other functions throughout the package.

It does not perform any additional cleanup beyond the separation and
does not have any functional changes.