txscript: optimize IsUnspendable by removing all allocations
During my work to optimize the IBD of btcd, I've been profiling heavily.
I've noticed that this call `IsUnspendable` shows up in several hot
spots during block validation, as we need to know if we should add a
UTXO to the view/cache, etc. The current version uses `parseScript` to
parse out all the pops and such. This is nice when we want to be able to
easily manipulate things like data pushes, but for this simple check,
it's unnecessary. The new version simply makes checks against the raw
script which saves us a ton of allocations and CPU cycles.