txscript/script: add zero-alloc countScriptSigOps
This commit adds a helper method
performing zero-alloc sigop counts.
We build on the prior changes by now
parsing the script in countOnly mode,
and returning the number of sigs.

The helper is then used in various
places where the parse opcodes were
simply discarded after getting the
sigop count.