btcec: remove import of chainhash
The chainhash import is challenging for consumers of the latest tagged
btcd module (v0.22.0-beta) because that module revision has its own
chainhash package.  This creates an ambiguous import when trying
to import both the older btcd module and btcec/v2 which requires the
standalone module version of chainhash.

This is the brute solution.  An alternative would be to create a
chaincfg/chainhash/v2 module and have btcec/v2 require that instead.