build: switch to using Go 1.18 for CI in go.mod
In this commit, we switch to using Go 1.18 everywhere. This is a
deviation from our typical "last 2 versions" rule. Switching now gives
us access to the new type params feature immediately, versus near the
end of this year. With this change we'd also be able to start using the
new std lib fuzzing features as well.