config: allow reading externalIPs from files
Tor hidden services dumps the onion hostname into a file on disk. If Tor
and btcd is started together -- for example in a sidecar pattern -- the
onion address is not known before executions.

This allows the user to provide --externalip a path to a file containing
any hostname, for example the onion address as dumped by the Tor
service.