GOGC and GOMAXPROCS configuration items added and defaults set
I think GOGC can be set via environment variable, the runtime checks
for it, maybe, but for certain it is better done directly in the app.

Increasing GOMAXPROCS above the number of CPU threads can produce
substantial improvements in initial block download and validation.

These both default to prior defaults if unset but provide a way for
deployments to dramatically improve IBD and speed of updating the DB
when new blocks arrive. Not important improvements but produce a ~50%
time to initial sync at the cost of around double memory use.

IBD is a definite weak point for btcd versus bitcoind. There is
minimal cost efficiency between a 4gb vs 2gb memory availability
nowadays, so this brings btcd closer to parity of performance.