txscript: add support for miniscript in the P2WSH context
Currently, only the P2WSH context is supported. Miniscript for Taproot
was not yet well specified at the time of writing. Currently, there is
an open PR for Bitcoin Core at
https://github.com/bitcoin/bitcoin/pull/27255 which, when merged, can
be taken as a reference.

This implementation contains:

- miniscript parser
- type checker
- malleability check
- script length limit check (P2WSH standardness rule)
- op count check (P2WSH consensus rule)
- duplicate pubkey check
- script generation
- witness generation (satisfactions)
- unit tests that create receive addresses based on some miniscripts and spend from them again (see
`TestRedeem()`).

Unit tests exist to check that this implementation's type checking passes some of the unit tests of
rust-miniscript. See [./testdata](./testdata).

Missing/todo:
- timelock mixing check
- stack size check (P2WSH standardness rule)
- witness generation for thresh/multi currently has a naive implementation that is very inefficient
(exponential runtime). This should be replaced with a fast algorithm.
- maybe more

Co-authored-by: Oliver Gugger <gugger@gmail.com>