netsync: ignore tx invs when we're not current
During ibd sometimes we may ban our sync peers because we're attempting
to fetch some txs that may already been confirmed and don't exist in the
mempool anymore.

Since we shouldn't be fetching txs anyways when we're not current, this
commit changes the handleInvMsg logic to ignore tx invs if the node is
not caught up to the tip.

Fixes #1984