netsync: don't reset headersFirstMode on resetHeaderState
We only exit headersFirstMode when we have no more checkpoints left.
updateSyncPeer was resetting headersFirstMode by calling
resetHeaderState only to have it turned back on with startSync().

Resetting the headers-first mode when we've not really exited the
headers first mode will cause problems with later commits that add
parallel block downloads so we remove it here.