netsync: change handleBlockMsgHeadersFirst to account for the fetchManager based download
Since we can't keep track of which block we've requested to which peer,
that we don't bother removing it from the peerstate.

Also, we no longer fetch more blocks based off of minInFlightBlocks.
We check if all the blocks in the window have been downloaded and ask
for more if they have been.